BCM2711: Allow more than one device on pcie bus >1
There is only a single pcie port on the bcm2711 so limiting the number of
devices to 1 worked as long as there is no way to add a pcie switch.

On the compute module 4 it is possible to add a pcie switch (tested with
asm1184e) which adds 5 new pcie busses.

In the current state the pci enumeration fails for the pcie switch
internal bus (bus 2, device 1,3,5,7). The root port gets configured with
subordniate=0x2, blocking e.g. a booting linux from discovering devices
behind the switch.

Devices behind the switch work after lifting the device limit on busses
other than 0 and 1.